Revert "shift active server listener to home/index"
This reverts commit 75e26faadd8c4359dd4d02c30375c03fadcc91bd.
